+++ Modules/Wrappers/ApplicationEngine/include/otbWrapperParameter.h
@@ -146,12 +146,12 @@ public:
    * to provide functions like SetParameterInt, GetParameterString, etc.
    */
   virtual int                      ToInt() const;
-  virtual float                    ToFloat() const;
+  virtual double                   ToFloat() const;
   virtual std::string              ToString() const;
   virtual std::vector<std::string> ToStringList() const;
 
   virtual void FromInt(int);
-  virtual void FromFloat(float);
+  virtual void FromFloat(double);
   virtual void FromString(const std::string&);
   virtual void FromStringList(const std::vector<std::string>&);
